FIX: NFF face winding bug (actually it was in StandardShapes.cpp). FIX: StandardShapes::MakeCone() - face order was incorrect in 50% of all cases. Implemented StandardShapes::MakeCircle(). Fixed a compiler warning in B3DImporter.cpp. Modified cone.nff - one texture isn't found, that's ok. The mapping should be better visible now.
git-svn-id: https://assimp.svn.sourceforge.net/svnroot/assimp/trunk@299 67173fc5-114c-0410-ac8e-9d2fd5bffc1f

+++ code/B3DImporter.cpp
@@ -247,7 +247,7 @@ void B3DImporter::ReadBRUS(){
 		//Textures
 		for( int i=0;i<n_texs;++i ){
 			int texid=ReadInt();
-			if( !i && texid>=0 && texid<_textures.size() ){
+			if( !i && texid>=0 && texid<(int)_textures.size() ){
 				//just use tex 0 for now
 				const Texture &tex=_textures[texid];
 				aiString texname( tex.name );
